+++ packages/utils/json_schema/lib/json_schema.dart
@@ -3,6 +3,7 @@ library;
 
 export 'src/from_schema.dart';
 export 'src/json_encodable.dart';
+export 'src/json_pointer.dart';
 export 'src/json_schema.dart';
 export 'src/json_schema_ref.dart';
 export 'src/json_schema_registry.dart';